+++ library/src/test/java/org/mustangproject/ZUGFeRD/IZUGFeRDExportableTransactionImpl.java
@@ -32,7 +32,6 @@ public class IZUGFeRDExportableTransactionImpl implements IZUGFeRDExportableTran
 	private IZUGFeRDExportableItem[] zFItems;
 	private IZUGFeRDExportableContact recipient;
 	private IZUGFeRDTradeSettlementPayment[] settlementPayments;
-	private IZUGFeRDDate zfDeliveryDate;
 	private String ownTaxID;
 	private String ownVATID;
 	private String ownOrganisationName;
@@ -150,10 +149,6 @@ public class IZUGFeRDExportableTransactionImpl implements IZUGFeRDExportableTran
 		return referenceNumber;
 	}
 
-	@Override
-	public IZUGFeRDDate getZFDeliveryDate() {
-		return zfDeliveryDate;
-	}
 
 	public IZUGFeRDExportableTransactionImpl setNumber(String number) {
 		this.number = number;
@@ -239,12 +234,7 @@ public class IZUGFeRDExportableTransactionImpl implements IZUGFeRDExportableTran
 		this.deliveryDate = deliveryDate;
 		return this;
 	}
-
-	public IZUGFeRDExportableTransactionImpl setZfDeliveryDate(IZUGFeRDDate zfDeliveryDate) {
-		this.zfDeliveryDate = zfDeliveryDate;
-		return this;
-	}
-
+	
 	public IZUGFeRDExportableTransactionImpl setCurrency(String currency) {
 		this.currency = currency;
 		return this;