+++ History.md
@@ -26,7 +26,9 @@ done
 - remove migration
 - be able to recursively validate directories
 - be able to ignore input pdf errors with -i
-
+- upgraded to verapdf 1.16.1
+- added Ghostscript signature
+- removal of izugferddate/IZUGFeRDPaymentTerms,IZUGFeRDPaymentDiscountTerms
 1.7.7
 =====
 2020-05-26